requirements.txt
python-binance
rich
typer
pydantic
python-dotenv
questionary
